order-service/requirements.txt
flask
mysql-connector-python
kafka-python
